+++ apps/web-app/src/pages/demo.tsx
@@ -1,0 +1,22 @@
+import { GetServerSideProps } from 'next';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import { BadRequest } from '@tsed/exceptions';
+import { DemoPage } from '../features/demo/demo.page';
+import { demoConfig } from '../features/demo/demo.config';
+
+export default function DemoRoute() {
+  return <DemoPage />;
+}
+
+export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
+  if (locale === undefined) {
+    throw new BadRequest('locale is missing');
+  }
+  const { i18nNamespaces } = demoConfig;
+  return {
+    props: {
+      // @see https:/github.com/i18next/react-i18next/pull/1340#issuecomment-874728587
+      ...(await serverSideTranslations(locale, i18nNamespaces.slice())),
+    },
+  };
+};